--- src/main/java/blackjack/model/GameSummary.java
@@ -1,4 +1,0 @@
-package blackjack.model;
-
-public record GameSummary(User user, int totalScore, boolean bust, boolean blackjack) {
-}
