+++ src/main/java/blackjack/model/PlayersGameResult.java
@@ -1,5 +1,6 @@
 package blackjack.model;
 
+import blackjack.model.user.Player;
 import java.util.Map;
 
 public record PlayersGameResult(Map<Player, GameResult> result) {
