+++ src/main/java/blackjack/model/card/Card.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 public class Card {
