+++ src/main/java/blackjack/model/constant/Constant.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.constant;
 
 public class Constant {
