+++ src/main/java/blackjack/model/constant/ErrorMessage.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.constant;
 
 public enum ErrorMessage {
