+++ src/main/java/blackjack/model/user/Player.java
@@ -1,5 +1,7 @@
-package blackjack.model;
+package blackjack.model.user;
 
+
+import blackjack.model.constant.Constant;
 
 public class Player extends User {
     public Player(String name) {
