+++ src/test/java/blackjack/util/PlayerParserTest.java
@@ -3,7 +3,7 @@ package blackjack.util;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.model.Player;
+import blackjack.model.user.Player;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
